+++ components/led/led_indicator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ChangeLog
 
+## v0.6.0 - 2023-10-24
+
+### Enhancements:
+
+* Support LED RGB type by driver [led_strip](https://components.espressif.com/components/espressif/led_strip)
+
+* Add api to control led indicator's brightness, HSV color, RGB color in real time
+
+* Support change led indicator's color by type HSV and RGB.
+
+* Support color gradient modification based on the color ring.
+
 ## v0.5.1 - 2023-10-23
 
 ### BUG FIX